Reused get_semantic_scholar_paper_id_by_article_doi
